requirements.txt
aiohttp==3.10.10
anyio==3.7.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2023.7.22
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
google-play-scraper==1.2.7
gplaycli==3.29
idna==3.4
joblib==1.4.2
langdetect==1.0.9
lxml==5.3.0
matlink-gpapi==0.4.4.5
numpy==2.1.2
openai==0.28.0
packaging==24.1
protobuf==5.28.3
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.25.0
six==1.16.0
soupsieve==2.5
tabulate==0.9.0
threadpoolctl==3.5.0
tqdm==4.66.6
typing_extensions==4.12.2
webdriver-manager==4.0.2
websocket-client==1.8.0

flask~=3.0.3
flask-cors~=5.0.0
firebase-admin~=6.5.0